🎨 라우트 구조 변경
