+++ frontend/src/pages/category/Pages1.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const Pages1 = () => {
-  return <div>첫번째 페이지</div>;
+  return <div>첫번째 페이지ㅁㅁ</div>;
 };
 
 export default Pages1;
