+++ frontend/src/pages/home/Home.module.css
@@ -1,0 +1,7 @@
+.box {
+  width: 200px;
+  height: 200px;
+  background-color: aqua;
+
+  margin-top: 20px;
+}
